+++ extensions/gleam-packages/src/components/PackageListItem.tsx
@@ -10,7 +10,7 @@ type PackageListItemProps = {
 export default function PackageListItem({ pkg, refreshAction }: PackageListItemProps) {
   return (
     <List.Item
-      key={pkg.id}
+      key={pkg.name}
       title={pkg.name}
       subtitle={pkg.description}
       keywords={keywords(pkg.description)}
@@ -19,13 +19,13 @@ export default function PackageListItem({ pkg, refreshAction }: PackageListItemP
           <Action.Push
             title="Show Package"
             icon={Icon.AppWindowSidebarLeft}
-            target={<DocsList packageName={pkg.name} url={pkg.docs_url} />}
+            target={<DocsList packageName={pkg.name} url={`https://hexdocs.pm/${pkg.name}/`} />}
           />
-          <Action.OpenInBrowser title="Open in Browser" url={pkg.docs_url} />
+          <Action.OpenInBrowser title="Open in Browser" url={`https://hexdocs.pm/${pkg.name}/`} />
           <Action.CopyToClipboard
             title="Copy Link"
             shortcut={{ modifiers: ["cmd", "shift"], key: "c" }}
-            content={pkg.docs_url}
+            content={`https://hexdocs.pm/${pkg.name}/`}
           />
           <ActionPanel.Section>{refreshAction}</ActionPanel.Section>
         </ActionPanel>