+++ extensions/gleam-packages/CHANGELOG.md
@@ -1,3 +1,6 @@
 # Gleam Packages Changelog
 
+## [Fix] - {PR_MERGE_DATE}
+- Uses the packages.gleam.run API to fetch package information.
+
 ## [Initial Version] - 2024-11-09